splitting up encoding decoding logic
seperating the protobin and xml encoding decoding will make it easier to maintain them
