+++ xml_converter/src/packaging_xml.cpp
@@ -1,0 +1,180 @@
+#include "packaging_xml.hpp"
+
+#include <utility>
+
+#include "rapid_helpers.hpp"
+#include "rapidxml-1.13/rapidxml.hpp"
+#include "rapidxml-1.13/rapidxml_utils.hpp"
+#include "string_helper.hpp"
+
+using namespace std;
+
+////////////////////////////////////////////////////////////////////////////////
+////////////////////////////////// SERIALIZE ///////////////////////////////////
+////////////////////////////////////////////////////////////////////////////////
+
+void parse_marker_categories(rapidxml::xml_node<>* node, map<string, Category>* marker_categories, vector<XMLError*>* errors, int depth = 0) {
+    if (get_node_name(node) == "MarkerCategory") {
+        string name = lowercase(find_attribute_value(node, "name"));
+
+        Category* this_category = &(*marker_categories)[name];
+        this_category->init_from_xml(node, errors);
+        for (rapidxml::xml_node<>* child_node = node->first_node(); child_node; child_node = child_node->next_sibling()) {
+            parse_marker_categories(child_node, &(this_category->children), errors, depth + 1);
+        }
+    }
+    else {
+        errors->push_back(new XMLNodeNameError("Unknown MarkerCategory Tag", node));
+    }
+}
+
+Category* get_category(rapidxml::xml_node<>* node, map<string, Category>* marker_categories, vector<XMLError*>* errors) {
+    // TODO: This is a slow linear search, replace with something faster.
+    //       maybe use data from already parsed node instead of searching for
+    //       the attribute.
+    rapidxml::xml_attribute<>* attribute = find_attribute(node, "type");
+
+    if (attribute == 0) {
+        // TODO: This error should really be for the entire node not just the name
+        errors->push_back(new XMLNodeNameError("No Attribute Named Type", node));
+        return nullptr;
+    }
+
+    vector<string> split_categories = split(get_attribute_value(attribute), ".");
+
+    if (split_categories.size() == 0) {
+        errors->push_back(new XMLAttributeValueError("Empty Type", attribute));
+        return nullptr;
+    }
+
+    Category* output = nullptr;
+
+    for (unsigned int i = 0; i < split_categories.size(); i++) {
+        string category_name = lowercase(split_categories[i]);
+
+        auto category = marker_categories->find(category_name);
+
+        if (category == marker_categories->end()) {
+            errors->push_back(new XMLAttributeValueError("Category Not Found \"" + category_name + "\"", attribute));
+            return nullptr;
+        }
+
+        output = &category->second;
+
+        marker_categories = &output->children;
+    }
+    return output;
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// parse_pois
+//
+// Parse the <POIs> xml block into an in-memory array of Markers.
+////////////////////////////////////////////////////////////////////////////////
+vector<Parseable*> parse_pois(rapidxml::xml_node<>* root_node, map<string, Category>* marker_categories, vector<XMLError*>* errors, string base_dir) {
+    vector<Parseable*> markers;
+
+    for (rapidxml::xml_node<>* node = root_node->first_node(); node; node = node->next_sibling()) {
+        if (get_node_name(node) == "POI") {
+            Category* default_category = get_category(node, marker_categories, errors);
+
+            Icon* icon = new Icon();
+
+            if (default_category != nullptr) {
+                *icon = default_category->default_icon;
+            }
+
+            icon->init_from_xml(node, errors);
+            markers.push_back(icon);
+        }
+        else if (get_node_name(node) == "Trail") {
+            Category* default_category = get_category(node, marker_categories, errors);
+
+            Trail* trail = new Trail();
+
+            if (default_category != nullptr) {
+                *trail = default_category->default_trail;
+            }
+
+            trail->init_from_xml(node, errors, base_dir);
+            markers.push_back(trail);
+        }
+        else {
+            errors->push_back(new XMLNodeNameError("Unknown POIs node name", node));
+        }
+    }
+    return markers;
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// parse_xml_file
+//
+// A function which parses a single XML file into their corrisponding classes.
+////////////////////////////////////////////////////////////////////////////////
+void parse_xml_file(string xml_filepath, map<string, Category>* marker_categories, vector<Parseable*>* parsed_pois) {
+    vector<XMLError*> errors;
+    rapidxml::xml_document<> doc;
+    rapidxml::xml_node<>* root_node;
+
+    rapidxml::file<> xml_file(xml_filepath.c_str());
+    doc.parse<rapidxml::parse_non_destructive | rapidxml::parse_no_data_nodes>(xml_file.data(), xml_filepath.c_str());
+
+    root_node = doc.first_node();
+    string base_dir = get_base_dir(xml_filepath);
+    // Validate the Root Node
+    if (get_node_name(root_node) != "OverlayData") {
+        errors.push_back(new XMLNodeNameError("Root node should be of type OverlayData", root_node));
+    }
+    if (root_node->first_attribute() != nullptr) {
+        cout << "Root Node has attributes when it should have none in " << xml_filepath << endl;
+    }
+
+    for (rapidxml::xml_node<>* node = root_node->first_node(); node; node = node->next_sibling()) {
+        if (get_node_name(node) == "MarkerCategory") {
+            parse_marker_categories(node, marker_categories, &errors);
+        }
+        else if (get_node_name(node) == "POIs") {
+            vector<Parseable*> temp_vector = parse_pois(node, marker_categories, &errors, base_dir);
+            move(temp_vector.begin(), temp_vector.end(), back_inserter(*parsed_pois));
+        }
+        else {
+            errors.push_back(new XMLNodeNameError("Unknown top-level node name", node));
+        }
+    }
+
+    for (auto error : errors) {
+        error->print_error();
+    }
+}
+
+////////////////////////////////////////////////////////////////////////////////
+///////////////////////////////// DESERIALIZE //////////////////////////////////
+////////////////////////////////////////////////////////////////////////////////
+
+void write_xml_file(string xml_filepath, map<string, Category>* marker_categories, vector<Parseable*>* parsed_pois) {
+    ofstream outfile;
+    string tab_string;
+
+    outfile.open(xml_filepath, ios::out);
+
+    outfile << "<OverlayData>\n";
+    for (const auto& category : *marker_categories) {
+        string text;
+        for (const auto& s : category.second.as_xml()) {
+            text += s;
+        }
+        outfile << text + "\n";
+    }
+
+    outfile << "<POIs>\n";
+    for (const auto& parsed_poi : *parsed_pois) {
+        string text;
+        for (const auto& s : parsed_poi->as_xml()) {
+            text += s;
+        }
+        outfile << text + "\n";
+    }
+    outfile << "</POIs>\n</OverlayData>\n";
+
+    outfile.close();
+}
