+++ xml_converter/src/packaging_xml.hpp
@@ -1,0 +1,20 @@
+#pragma once
+
+#include <map>
+#include <string>
+#include <vector>
+
+#include "category_gen.hpp"
+#include "parseable.hpp"
+#include "rapidxml-1.13/rapidxml.hpp"
+#include "rapidxml-1.13/rapidxml_utils.hpp"
+
+void parse_xml_file(
+    std::string xml_filepath,
+    std::map<std::string, Category>* marker_categories,
+    std::vector<Parseable*>* parsed_pois);
+
+void write_xml_file(
+    std::string xml_filepath,
+    std::map<std::string, Category>* marker_categories,
+    std::vector<Parseable*>* parsed_pois);
